+++ src/plugins/sql.ts
@@ -328,7 +328,7 @@ export class sql{
         return;
     }
 
-    calculateAnnualQuota(user:string,year:digit):{quota:number,years:number,month:number,days:number}{
+    calculateAnnualQuota(user:string,year:digit):{quota:number,years:number,month:number,days:number,joinTime:string}{
         const db_jt:string = (this.login_db.prepare(`SELECT * FROM userinfo WHERE id='${user}'`).all()[0] as userinfo)["joinTime"];
         const joinTime:Date = new Date(db_jt);
         const endTime:Date = new Date(`${year}-${db_jt.split("-")[1]}-${db_jt.split("-")[2]}`);
@@ -351,7 +351,7 @@ export class sql{
             const w = Math.floor(years)+6;
             quota = (w>=30?30:w);
         }
-        return {"quota":quota*8,"years":Math.floor(years),"month":(months-(Math.floor(years)*12)),"days":days};
+        return {"quota":quota*8,"years":Math.floor(years),"month":(months-(Math.floor(years)*12)),"days":days,"joinTime":db_jt};
 
     }