+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
+## [1.2.0] 2025-03-19
+
+### Changed
+- /approved now accept month parameter.
+- /quota now returns join time.
+
+---
 ## [1.1.0] 2025-03-19
 
 ### Changed